requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-image>=0.21.0
scipy>=1.10.0
jupyter>=1.0.0
